Add the possibility to export conversations as zip (xml+mediadata+viewer)
Add the possibility to export conversations as zip-file containing the
conversation in xml format. It is possible to define the desired period of time.
Mediadata can be included in the export.
An HTML viewer can also be added to view the exported chat.

This feature was developed as the subject of Patrick Nolte's bachelor
thesis at the Institute of Networks and Security at JKU.